requirements.txt
imageio==2.37.0
Pillow==11.3.0
python_fasthtml==0.12.23
